got server.js running
